Codewars Maximum Triplet Sum (Array Series #7) - 7kyu
Exercise that aims to check the three largest integers within an array, and not use repeated values. In order to find the 3 largest integer values, add them and result in a value. As logic, a For function and logic was used to define the three largest values.

Follow the link below for a better view:
https://www.codewars.com/kata/5aa1bcda373c2eb596000112